+++ assets/js/views/loader.js
@@ -1,17 +1,17 @@
 import MainView from './main';
+import AdminUsersView from './admin/users';
 import TopicEditView from './topic/edit';
 import TopicIndexView from './topic/index';
 import TopicNewView from './topic/new';
 import TopicShowView from './topic/show';
-import UserIndexView from './user/index';
 
 // Collection of specific view modules
 const views = {
+  AdminUsersView,
   TopicShowView,
   TopicEditView,
   TopicIndexView,
-  TopicNewView,
-  UserIndexView
+  TopicNewView
 };
 
 export default function loadView(viewName) {