+++ lib/collaboration/coherence/user.ex
@@ -10,10 +10,11 @@ defmodule Collaboration.Coherence.User do
     field :admin, :boolean, default: false
     field :name, :string
     field :email, :string
-    field :feedback, :boolean
+    field :feedback, :boolean, default: false
     coherence_schema()
     timestamps()
     has_many :comments, Collaboration.Contributions.Comment, on_delete: :delete_all
+    has_many :feedbacks, Collaboration.Contributions.Comment, on_delete: :delete_all
     has_many :ideas, Collaboration.Contributions.Idea, on_delete: :delete_all
     has_many :ratings, Collaboration.Contributions.Rating, on_delete: :delete_all
     many_to_many :likes, Collaboration.Contributions.Comment, join_through: "likes", on_delete: :delete_all