+++ priv/repo/migrations/20180422161609_comments_add_recipient.exs
@@ -1,0 +1,9 @@
+defmodule Collaboration.Repo.Migrations.CommentsAddRecipient do
+  use Ecto.Migration
+
+  def change do
+    alter table(:comments) do
+      add :recipient_id, references(:users, type: :binary_id, on_delete: :delete_all)
+    end
+  end
+end